Move PremiseFailure exception into premises package
