+++ org.metaborg.meta.lang.dynsem.interpreter/src/main/java/org/metaborg/meta/lang/dynsem/interpreter/nodes/rules/premises/PremiseFailure.java
@@ -1,4 +1,4 @@
-package org.metaborg.meta.lang.dynsem.interpreter;
+package org.metaborg.meta.lang.dynsem.interpreter.nodes.rules.premises;
 
 import com.oracle.truffle.api.nodes.ControlFlowException;
